Split single form in a form per step to prevent states overwriting other formState
